Refactor: Migrate Academic Staff pages to TypeScript, fix Specialization creation 500 error, and improve search UI
