+++ frontend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/academic-staff/academicstaffsidebar.tsx","./src/components/admin/adminheader.tsx","./src/components/admin/adminlayout.tsx","./src/components/admin/adminsidebar.tsx","./src/components/admin/dashboard/alertssection.tsx","./src/components/admin/dashboard/notificationssection.tsx","./src/components/admin/dashboard/recentaccesstable.tsx","./src/components/admin/dashboard/statcard.tsx","./src/components/admin/dashboard/systemlogssection.tsx","./src/components/admin/dashboard/vietnammap.tsx","./src/components/admin/users/bulkactions.tsx","./src/components/admin/users/userfilters.tsx","./src/components/admin/users/usermodals.tsx","./src/components/admin/users/usertablerow.tsx","./src/components/common/button.tsx","./src/components/common/card.tsx","./src/components/common/commonheader.tsx","./src/components/common/confirmmodal.tsx","./src/components/common/protectedroute.tsx","./src/components/common/publicroute.tsx","./src/config/colors.ts","./src/hooks/usewebsocket.ts","./src/layouts/academicstafflayout.tsx","./src/pages/academic-staff/academicstaffdashboard.tsx","./src/pages/academic-staff/majormanagement.jsx","./src/pages/academic-staff/profilepage.tsx","./src/pages/admin/activateduserspage.tsx","./src/pages/admin/alertspage.tsx","./src/pages/admin/dashboard.tsx","./src/pages/admin/lockeduserspage.tsx","./src/pages/admin/notificationspage.tsx","./src/pages/admin/profilepage.tsx","./src/pages/admin/recentaccesspage.tsx","./src/pages/admin/systemlogspage.tsx","./src/pages/admin/userspage.tsx","./src/pages/auth/changepasswordpage.tsx","./src/pages/auth/forgotpasswordpage.tsx","./src/pages/auth/login.tsx","./src/services/api/academicstaffservice.ts","./src/services/api/authservice.ts","./src/services/api/config.ts","./src/services/api/dashboardservice.ts","./src/services/api/majorservice.ts","./src/services/api/mapservice.ts","./src/services/api/userservice.ts","./src/types/dashboard.ts","./src/types/major.ts"],"version":"5.9.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/academic-staff/academicstaffsidebar.tsx","./src/components/academic-staff/pagination.tsx","./src/components/academic-staff/selectionactionbar.tsx","./src/components/academic-staff/statusbadge.tsx","./src/components/academic-staff/statusfilter.tsx","./src/components/academic-staff/index.ts","./src/components/admin/adminheader.tsx","./src/components/admin/adminlayout.tsx","./src/components/admin/adminsidebar.tsx","./src/components/admin/dashboard/alertssection.tsx","./src/components/admin/dashboard/notificationssection.tsx","./src/components/admin/dashboard/recentaccesstable.tsx","./src/components/admin/dashboard/statcard.tsx","./src/components/admin/dashboard/systemlogssection.tsx","./src/components/admin/dashboard/vietnammap.tsx","./src/components/admin/users/bulkactions.tsx","./src/components/admin/users/userfilters.tsx","./src/components/admin/users/usermodals.tsx","./src/components/admin/users/usertablerow.tsx","./src/components/common/button.tsx","./src/components/common/card.tsx","./src/components/common/commonheader.tsx","./src/components/common/confirmmodal.tsx","./src/components/common/protectedroute.tsx","./src/components/common/publicroute.tsx","./src/config/colors.ts","./src/hooks/usedebounce.js","./src/hooks/usewebsocket.ts","./src/layouts/academicstafflayout.tsx","./src/pages/academic-staff/academicstaffdashboard.tsx","./src/pages/academic-staff/majordetail.tsx","./src/pages/academic-staff/majormanagement.tsx","./src/pages/academic-staff/profilepage.tsx","./src/pages/admin/activateduserspage.tsx","./src/pages/admin/alertspage.tsx","./src/pages/admin/dashboard.tsx","./src/pages/admin/lockeduserspage.tsx","./src/pages/admin/notificationspage.tsx","./src/pages/admin/profilepage.tsx","./src/pages/admin/recentaccesspage.tsx","./src/pages/admin/systemlogspage.tsx","./src/pages/admin/userspage.tsx","./src/pages/auth/changepasswordpage.tsx","./src/pages/auth/forgotpasswordpage.tsx","./src/pages/auth/login.tsx","./src/services/api/academicstaffservice.ts","./src/services/api/authservice.ts","./src/services/api/config.ts","./src/services/api/dashboardservice.ts","./src/services/api/majorservice.ts","./src/services/api/mapservice.ts","./src/services/api/specializationservice.ts","./src/services/api/userservice.ts","./src/types/dashboard.ts","./src/types/major.ts","./src/types/specialization.ts"],"version":"5.9.3"}
